+++ src/molecules/collapse/Collapse.stories.js
@@ -1,21 +1,10 @@
 import { storiesOf } from '@storybook/vue'
-import { select, text } from '@storybook/addon-knobs'
-
-import generateVueInfoTable from '@utils/helpers/generate-vue-info-table.js'
-import getClassKnobsConfig from '@utils/helpers/get-class-knobs-config.js'
-import selectorsConfig from './Collapse.selectors.json'
+import { text } from '@storybook/addon-knobs'
 
 import ACollapse from './Collapse.vue'
 
-const info = `
-  <p>Check <b>Knobs</b> tab to edit component properties dynamically.</p><br>
-  ${generateVueInfoTable(selectorsConfig, 'BEM modifiers')}
-`
-
-const classKnobsConfig = getClassKnobsConfig(selectorsConfig)
-
 storiesOf('Molecules/Collapse', module)
-  .addParameters({ info })
+  .addParameters({ info: true })
   .add(
     'Default',
     () => ({
@@ -23,9 +12,6 @@ storiesOf('Molecules/Collapse', module)
         ACollapse
       },
       props: {
-        classKnobs: {
-          default: select('BEM Modifier', classKnobsConfig)
-        },
         titleTextKnobs: {
           default: text('Title', 'Collapse title text')
         },
@@ -34,10 +20,7 @@ storiesOf('Molecules/Collapse', module)
         }
       },
       template: `
-        <a-collapse
-          :class="classKnobs"
-          :title="titleTextKnobs"
-        >
+        <a-collapse :title="titleTextKnobs">
           <span>{{ contentTextKnobs }}</span>
         </a-collapse>
       `